+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Sort env variables by name
 - Update Vue to v3.4.38
 - Update Uppy to 4.2.0
+- Fix late fields projection used on populated fields
 
 ## 2.9.0 (2024-08-08)